@@ -1,108 +1,100 @@
-import apigateway = require('@aws-cdk/aws-apigateway'); 
-import dynamodb = require('@aws-cdk/aws-dynamodb');
-import lambda = require('@aws-cdk/aws-lambda');
-import cdk = require('@aws-cdk/core');
+import { IResource, LambdaIntegration, MockIntegration, PassthroughBehavior, RestApi } from '@aws-cdk/aws-apigateway';
+import { AttributeType, Table } from '@aws-cdk/aws-dynamodb';
+import { Runtime } from '@aws-cdk/aws-lambda';
+import { App, Stack, RemovalPolicy } from '@aws-cdk/core';
+import { NodejsFunction, NodejsFunctionProps } from '@aws-cdk/aws-lambda-nodejs';
+import { join } from 'path'
 
-export class ApiLambdaCrudDynamoDBStack extends cdk.Stack {
-  constructor(app: cdk.App, id: string) {
+export class ApiLambdaCrudDynamoDBStack extends Stack {
+  constructor(app: App, id: string) {
     super(app, id);
 
-    const dynamoTable = new dynamodb.Table(this, 'items', {
+    const dynamoTable = new Table(this, 'items', {
       partitionKey: {
         name: 'itemId',
-        type: dynamodb.AttributeType.STRING
+        type: AttributeType.STRING
       },
       tableName: 'items',
 
-      // The default removal policy is RETAIN, which means that cdk destroy will not attempt to delete
-      // the new table, and it will remain in your account until manually deleted. By setting the policy to 
-      // DESTROY, cdk destroy will delete the table (even if it has data in it)
-      removalPolicy: cdk.RemovalPolicy.DESTROY, // NOT recommended for production code
+      /**
+       *  The default removal policy is RETAIN, which means that cdk destroy will not attempt to delete
+       * the new table, and it will remain in your account until manually deleted. By setting the policy to
+       * DESTROY, cdk destroy will delete the table (even if it has data in it)
+       */
+      removalPolicy: RemovalPolicy.DESTROY, // NOT recommended for production code
     });
 
-    const getOneLambda = new lambda.Function(this, 'getOneItemFunction', {
-      code: new lambda.AssetCode('src'),
-      handler: 'get-one.handler',
-      runtime: lambda.Runtime.NODEJS_10_X,
+    const nodeJsFunctionProps: NodejsFunctionProps = {
+      bundling: {
+        externalModules: [
+          'aws-sdk', // Use the 'aws-sdk' available in the Lambda runtime
+        ],
+      },
+      depsLockFilePath: join(__dirname, 'lambdas', 'package-lock.json'),
       environment: {
+        PRIMARY_KEY: 'itemId',
         TABLE_NAME: dynamoTable.tableName,
-        PRIMARY_KEY: 'itemId'
-      }
+      },
+      runtime: Runtime.NODEJS_14_X,
+    }
+
+    // Create a Lambda function for each of the CRUD operations
+    const getOneLambda = new NodejsFunction(this, 'getOneItemFunction', {
+      entry: join(__dirname, 'lambdas', 'get-one.ts'),
+      ...nodeJsFunctionProps,
+    });
+    const getAllLambda = new NodejsFunction(this, 'getAllItemsFunction', {
+      entry: join(__dirname, 'lambdas', 'get-all.ts'),
+      ...nodeJsFunctionProps,
+    });
+    const createOneLambda = new NodejsFunction(this, 'createItemFunction', {
+      entry: join(__dirname, 'lambdas', 'create.ts'),
+      ...nodeJsFunctionProps,
+    });
+    const updateOneLambda = new NodejsFunction(this, 'updateItemFunction', {
+      entry: join(__dirname, 'lambdas', 'update-one.ts'),
+      ...nodeJsFunctionProps,
+    });
+    const deleteOneLambda = new NodejsFunction(this, 'deleteItemFunction', {
+      entry: join(__dirname, 'lambdas', 'delete-one.ts'),
+      ...nodeJsFunctionProps,
     });
 
-    const getAllLambda = new lambda.Function(this, 'getAllItemsFunction', {
-      code: new lambda.AssetCode('src'),
-      handler: 'get-all.handler',
-      runtime: lambda.Runtime.NODEJS_10_X,
-      environment: {
-        TABLE_NAME: dynamoTable.tableName,
-        PRIMARY_KEY: 'itemId'
-      }
-    });
-
-    const createOne = new lambda.Function(this, 'createItemFunction', {
-      code: new lambda.AssetCode('src'),
-      handler: 'create.handler',
-      runtime: lambda.Runtime.NODEJS_10_X,
-      environment: {
-        TABLE_NAME: dynamoTable.tableName,
-        PRIMARY_KEY: 'itemId'
-      }
-    });
-
-    const updateOne = new lambda.Function(this, 'updateItemFunction', {
-      code: new lambda.AssetCode('src'),
-      handler: 'update-one.handler',
-      runtime: lambda.Runtime.NODEJS_10_X,
-      environment: {
-        TABLE_NAME: dynamoTable.tableName,
-        PRIMARY_KEY: 'itemId'
-      }
-    });
-
-    const deleteOne = new lambda.Function(this, 'deleteItemFunction', {
-      code: new lambda.AssetCode('src'),
-      handler: 'delete-one.handler',
-      runtime: lambda.Runtime.NODEJS_10_X,
-      environment: {
-        TABLE_NAME: dynamoTable.tableName,
-        PRIMARY_KEY: 'itemId'
-      }
-    });
-    
+    // Grant the Lambda function read access to the DynamoDB table
     dynamoTable.grantReadWriteData(getAllLambda);
     dynamoTable.grantReadWriteData(getOneLambda);
-    dynamoTable.grantReadWriteData(createOne);
-    dynamoTable.grantReadWriteData(updateOne);
-    dynamoTable.grantReadWriteData(deleteOne);
+    dynamoTable.grantReadWriteData(createOneLambda);
+    dynamoTable.grantReadWriteData(updateOneLambda);
+    dynamoTable.grantReadWriteData(deleteOneLambda);
 
-    const api = new apigateway.RestApi(this, 'itemsApi', {
+    // Integrate the Lambda functions with the API Gateway resource
+    const getAllIntegration = new LambdaIntegration(getAllLambda);
+    const createOneIntegration = new LambdaIntegration(createOneLambda);
+    const getOneIntegration = new LambdaIntegration(getOneLambda);
+    const updateOneIntegration = new LambdaIntegration(updateOneLambda);
+    const deleteOneIntegration = new LambdaIntegration(deleteOneLambda);
+
+
+    // Create an API Gateway resource for each of the CRUD operations
+    const api = new RestApi(this, 'itemsApi', {
       restApiName: 'Items Service'
     });
 
     const items = api.root.addResource('items');
-    const getAllIntegration = new apigateway.LambdaIntegration(getAllLambda);
     items.addMethod('GET', getAllIntegration);
-
-    const createOneIntegration = new apigateway.LambdaIntegration(createOne);
     items.addMethod('POST', createOneIntegration);
     addCorsOptions(items);
 
     const singleItem = items.addResource('{id}');
-    const getOneIntegration = new apigateway.LambdaIntegration(getOneLambda);
     singleItem.addMethod('GET', getOneIntegration);
-
-    const updateOneIntegration = new apigateway.LambdaIntegration(updateOne);
     singleItem.addMethod('PATCH', updateOneIntegration);
-
-    const deleteOneIntegration = new apigateway.LambdaIntegration(deleteOne);
     singleItem.addMethod('DELETE', deleteOneIntegration);
     addCorsOptions(singleItem);
   }
 }
 
-export function addCorsOptions(apiResource: apigateway.IResource) {
-  apiResource.addMethod('OPTIONS', new apigateway.MockIntegration({
+export function addCorsOptions(apiResource: IResource) {
+  apiResource.addMethod('OPTIONS', new MockIntegration({
     integrationResponses: [{
       statusCode: '200',
       responseParameters: {
@@ -112,7 +104,7 @@ export function addCorsOptions(apiResource: apigateway.IResource) {
         'method.response.header.Access-Control-Allow-Methods': "'OPTIONS,GET,PUT,POST,DELETE'",
       },
     }],
-    passthroughBehavior: apigateway.PassthroughBehavior.NEVER,
+    passthroughBehavior: PassthroughBehavior.NEVER,
     requestTemplates: {
       "application/json": "{\"statusCode\": 200}"
     },
@@ -124,11 +116,11 @@ export function addCorsOptions(apiResource: apigateway.IResource) {
         'method.response.header.Access-Control-Allow-Methods': true,
         'method.response.header.Access-Control-Allow-Credentials': true,
         'method.response.header.Access-Control-Allow-Origin': true,
-      },  
+      },
     }]
   })
 }
 
-const app = new cdk.App();
+const app = new App();
 new ApiLambdaCrudDynamoDBStack(app, 'ApiLambdaCrudDynamoDBExample');
 app.synth();
