@@ -1,13 +1,15 @@
-const AWS = require('aws-sdk');
-const db = new AWS.DynamoDB.DocumentClient();
-const uuidv4 = require('uuid/v4');
+import * as AWS from 'aws-sdk';
+import { v4 as uuidv4 } from 'uuid';
+
 const TABLE_NAME = process.env.TABLE_NAME || '';
 const PRIMARY_KEY = process.env.PRIMARY_KEY || '';
+
+const db = new AWS.DynamoDB.DocumentClient();
 
 const RESERVED_RESPONSE = `Error: You're using AWS reserved keywords as attributes`,
   DYNAMODB_EXECUTION_ERROR = `Error: Execution update, caused a Dynamodb error, please take a look at your CloudWatch Logs.`;
 
-export const handler = async (event: any = {}) : Promise <any> => {
+export const handler = async (event: any = {}): Promise<any> => {
 
   if (!event.body) {
     return { statusCode: 400, body: 'invalid request, you are missing the parameter body' };
@@ -24,7 +26,7 @@ export const handler = async (event: any = {}) : Promise <any> => {
     return { statusCode: 201, body: '' };
   } catch (dbError) {
     const errorResponse = dbError.code === 'ValidationException' && dbError.message.includes('reserved keyword') ?
-    DYNAMODB_EXECUTION_ERROR : RESERVED_RESPONSE;
+      DYNAMODB_EXECUTION_ERROR : RESERVED_RESPONSE;
     return { statusCode: 500, body: errorResponse };
   }
 };
