@@ -1,8 +1,10 @@
-const AWS = require('aws-sdk');
-const db = new AWS.DynamoDB.DocumentClient();
+import * as AWS from 'aws-sdk';
+
 const TABLE_NAME = process.env.TABLE_NAME || '';
 
-export const handler = async () : Promise <any> => {
+const db = new AWS.DynamoDB.DocumentClient();
+
+export const handler = async (): Promise<any> => {
 
   const params = {
     TableName: TABLE_NAME
@@ -12,6 +14,6 @@ export const handler = async () : Promise <any> => {
     const response = await db.scan(params).promise();
     return { statusCode: 200, body: JSON.stringify(response.Items) };
   } catch (dbError) {
-    return { statusCode: 500, body: JSON.stringify(dbError)};
+    return { statusCode: 500, body: JSON.stringify(dbError) };
   }
 };
