@@ -1,9 +1,11 @@
-const AWS = require('aws-sdk');
-const db = new AWS.DynamoDB.DocumentClient();
+import * as AWS from 'aws-sdk';
+
 const TABLE_NAME = process.env.TABLE_NAME || '';
 const PRIMARY_KEY = process.env.PRIMARY_KEY || '';
 
-export const handler = async (event: any = {}) : Promise <any> => {
+const db = new AWS.DynamoDB.DocumentClient();
+
+export const handler = async (event: any = {}): Promise<any> => {
 
   const requestedItemId = event.pathParameters.id;
   if (!requestedItemId) {
@@ -18,8 +20,8 @@ export const handler = async (event: any = {}) : Promise <any> => {
   };
 
   try {
-    await db.delete(params).promise();
-    return { statusCode: 200, body: '' };
+    const response = await db.get(params).promise();
+    return { statusCode: 200, body: JSON.stringify(response.Item) };
   } catch (dbError) {
     return { statusCode: 500, body: JSON.stringify(dbError) };
   }
