@@ -1,12 +1,15 @@
-const AWS = require('aws-sdk');
-const db = new AWS.DynamoDB.DocumentClient();
+import * as AWS from 'aws-sdk';
+
 const TABLE_NAME = process.env.TABLE_NAME || '';
 const PRIMARY_KEY = process.env.PRIMARY_KEY || '';
+
 
 const RESERVED_RESPONSE = `Error: You're using AWS reserved keywords as attributes`,
   DYNAMODB_EXECUTION_ERROR = `Error: Execution update, caused a Dynamodb error, please take a look at your CloudWatch Logs.`;
 
-export const handler = async (event: any = {}) : Promise <any> => {
+const db = new AWS.DynamoDB.DocumentClient();
+
+export const handler = async (event: any = {}): Promise<any> => {
 
   if (!event.body) {
     return { statusCode: 400, body: 'invalid request, you are missing the parameter body' };
@@ -20,24 +23,24 @@ export const handler = async (event: any = {}) : Promise <any> => {
   const editedItem: any = typeof event.body == 'object' ? event.body : JSON.parse(event.body);
   const editedItemProperties = Object.keys(editedItem);
   if (!editedItem || editedItemProperties.length < 1) {
-      return { statusCode: 400, body: 'invalid request, no arguments provided' };
+    return { statusCode: 400, body: 'invalid request, no arguments provided' };
   }
 
-  const firstProperty = editedItemProperties.splice(0,1);
+  const firstProperty = editedItemProperties.splice(0, 1);
   const params: any = {
-      TableName: TABLE_NAME,
-      Key: {
-        [PRIMARY_KEY]: editedItemId
-      },
-      UpdateExpression: `set ${firstProperty} = :${firstProperty}`,
-      ExpressionAttributeValues: {},
-      ReturnValues: 'UPDATED_NEW'
+    TableName: TABLE_NAME,
+    Key: {
+      [PRIMARY_KEY]: editedItemId
+    },
+    UpdateExpression: `set ${firstProperty} = :${firstProperty}`,
+    ExpressionAttributeValues: {},
+    ReturnValues: 'UPDATED_NEW'
   }
   params.ExpressionAttributeValues[`:${firstProperty}`] = editedItem[`${firstProperty}`];
 
   editedItemProperties.forEach(property => {
-      params.UpdateExpression += `, ${property} = :${property}`;
-      params.ExpressionAttributeValues[`:${property}`] = editedItem[property];
+    params.UpdateExpression += `, ${property} = :${property}`;
+    params.ExpressionAttributeValues[`:${property}`] = editedItem[property];
   });
 
   try {
@@ -45,7 +48,7 @@ export const handler = async (event: any = {}) : Promise <any> => {
     return { statusCode: 204, body: '' };
   } catch (dbError) {
     const errorResponse = dbError.code === 'ValidationException' && dbError.message.includes('reserved keyword') ?
-    DYNAMODB_EXECUTION_ERROR : RESERVED_RESPONSE;
+      DYNAMODB_EXECUTION_ERROR : RESERVED_RESPONSE;
     return { statusCode: 500, body: errorResponse };
   }
 };
